Add functionality for `do_predict_proba` to be an integer, as well as a boolean
This allows one to specify a particular column from a model's probability predictions to pass
forward as the final predictions. If `do_predict_proba` is an int, then the `predict` method of
`models.Model` will call `predict_proba` on the inner model to return probability predictions,
and the integer identifies the index of the column whose values should be used as predictions.
Original behavior when providing boolean values for `do_predict_proba` remains unchanged.