Perform miscellaneous minor refactoring and identify bug with tracking best experiment
Mark location in `BaseOptimizationProtocol._optimization_loop` at which there is a bug that
does not properly record `best_experiment`, or `best_score` when the `target_metric` should be minimized.

This bug does not affect the actual optimization routine or its results. It only results in the incorrect
best score and experiment id being printed to the console at the end of optimization