Add `TranslateTrace` function/method decorator class
This is the first, lightly tested version of `TranslateTrace` and should not
be used anywhere yet, except for testing purposes. The plan is to use
`TranslateTrace` to decorate methods that accept, for example, the following
kwargs: `model`, `model_initializer`, `model_init_params`. The decorator, if
used in conjunction with a `Mirror` (also in testing) will convert inputs to
`model` to `model_initializer`, and `model_init_params`, instead. This
behavior is documented in greater detail in `tracers.TranslateTrace`, and
will be further documented as it is developed.