Allow `handle_complex_types` to proceed when `tested_keys_dir=None`. Closes #114
Instead of returning from the method immediately when `tested_keys_dir=None`,
the hashing of complex types is allowed to continue. Only the portions
that actually save files to the hash lookup directory are skipped when
`tested_keys_dir=None`. Most notably, this enables proper handling of `Metric`
instances, closing #114