Decorate `add_complex_type_lookup_entry` with `RetryMakeDirs`
This situation differs from the other refactorings with `RetryMakeDirs`
in that `make_dirs` is still manually called in one case. If the complex-typed
value is going to be `shelve`d, `make_dirs` is called automatically because
`RetryMakeDirs` would break otherwise.

This is because if `RetryMakeDirs` catches an OSError it is not going to be a
standard OSError or FileNotFoundError. The exception most likely thrown in this
case is some weird `_dbm` descendant of OSError, which does not have the `filename`
attribute required by `RetryMakeDirs` set as expected. In order to avoid this,
`make_dirs` with `exist_ok=True` is always called prior to `shelve.open`